source sets fix 2
