--- app/src/regular/java/com/example/ppolab/MainActivity.java
@@ -1,16 +1,0 @@
-package com.example.ppolab;
-
-import android.os.Bundle;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-public class MainActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.main_activity);
-
-
-}
-}
