+++ app/src/regular/res/layout/input.xml
@@ -46,5 +46,10 @@
         android:orientation="horizontal">
 
     </LinearLayout>
+    <Button android:text="swap"
+        android:id="@+id/buttonswap"
 
+        android:layout_height="wrap_content"
+        android:layout_width="wrap_content"
+        android:visibility="invisible"/>
 </LinearLayout>
