+++ app/src/regular/res/layout/main_buttons.xml
@@ -50,7 +50,14 @@
         android:id="@+id/button9"
         style="@style/ButtonStyle"
         />
-    <TextView />
+    <Button android:text="copy_1"
+        android:id="@+id/copy1"
+        android:layout_height="45dp"
+        android:layout_width="90dp"
+        android:visibility="invisible"
+        />
+
+
     <Button android:text="@string/b"
         android:id="@+id/buttonb"
         style="@style/ButtonStyle"
@@ -64,6 +71,13 @@
         android:id="@+id/buttonc"
         style="@style/ButtonStyle"
         />
+    <Button android:text="copy_2"
+        android:id="@+id/copy2"
+        android:layout_height="45dp"
+        android:layout_width="90dp"
+        android:visibility="invisible"
+        />
+
